rename a few more files
